Further encapsulate allocation of slots
