+++ flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/declarative/allocator/RequirementsCalculator.java
@@ -15,26 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.scheduler.declarative;
+package org.apache.flink.runtime.scheduler.declarative.allocator;
 
-import org.apache.flink.runtime.jobmaster.SlotInfo;
+import org.apache.flink.runtime.jobgraph.JobVertex;
+import org.apache.flink.runtime.jobmaster.slotpool.ResourceCounter;
 
-/** An execution slot sharing group and the slot it could be assigned to. */
-public class ExecutionSlotSharingGroupAndSlot {
-    private final ExecutionSlotSharingGroup executionSlotSharingGroup;
-    private final SlotInfo slotInfo;
+/** Calculates resource requirements for a set of vertices. */
+interface RequirementsCalculator {
 
-    public ExecutionSlotSharingGroupAndSlot(
-            ExecutionSlotSharingGroup executionSlotSharingGroup, SlotInfo slotInfo) {
-        this.executionSlotSharingGroup = executionSlotSharingGroup;
-        this.slotInfo = slotInfo;
-    }
-
-    public ExecutionSlotSharingGroup getExecutionSlotSharingGroup() {
-        return executionSlotSharingGroup;
-    }
-
-    public SlotInfo getSlotInfo() {
-        return slotInfo;
-    }
+    /**
+     * Calculates the total resources required for scheduling the given vertices.
+     *
+     * @param vertices vertices to schedule
+     * @return required resources
+     */
+    // TODO: replace JobVertex with VertexInformation
+    ResourceCounter calculateRequiredSlots(Iterable<JobVertex> vertices);
 }
